+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Allows user to follow an artist from the Artwork tombstone - lilyfromseattle
 - Enable `strictBindCallApply`, `noImplicitThis`, and `strictFunctionTypes` compiler settings - alloy
 - Adds ReadMore component to work with markdown - ds300/sweir27
+- Upgrade react-spring@8.0.23 - damassi
 
 ### 1.11.4